Improved output, Asks for confirmation 2 proceed
Message IDS are listed with a number to quantify how many relvant
messages will be deleted after all the filtering is done.

Asks for a typed confirmation before being able to delete.

Deletion is not implemented yet, and won't be until the end of the
project, as I have to test this on my own live data and I must be
confident so as to not ruin my email or generate myself a task.